#15 - Flutter Markdown 사용해보기
이전에 사용하던 WYSIWYG 편집기인 `Flutter Quill` 라이브러리는 사용자가 사용하기에는 편하지만, 개발자 입장에서는 사용하기 까다로웠다.

또한, json 형식으로 저장되므로 원문 자체를 읽기에 어려울 수 있다.

따라서, `Flutter Markdown` 라이브러리를 사용하여 게시글을 작성할 때 markdown 문법을 사용하여 작성하면, 작성한 게시글을 Markdown 형식의 게시글을 보여주도록 하였다.

Markdown 을 사용하기 위해 러닝 커브가 존재하지만, 그다지 어렵지 않아 괜찮을 듯 싶다.

- https://www.markdownguide.org/
