+++ lib/screen/main/tab/board/screen/s_write_board.dart
@@ -1,15 +1,10 @@
-import 'dart:convert';
-
 import 'package:flutter/material.dart';
-import 'package:flutter_quill/flutter_quill.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:playground/screen/main/tab/board/f_board.riverpod.dart';
-import 'package:playground/screen/main/tab/board/vo/vo_board_list.dart';
 
 import '../vo/vo_board.dart';
 
 class WriteBoardScreen extends ConsumerStatefulWidget {
-
   const WriteBoardScreen({super.key});
 
   @override
@@ -17,7 +12,7 @@ class WriteBoardScreen extends ConsumerStatefulWidget {
 }
 
 class _WriteBoardScreenState extends ConsumerState<WriteBoardScreen> {
-  final QuillController _controller = QuillController.basic();
+  final TextEditingController contentController = TextEditingController();
   final TextEditingController _textEditingController = TextEditingController();
   final _formKey = GlobalKey<FormState>();
 
@@ -29,51 +24,51 @@ class _WriteBoardScreenState extends ConsumerState<WriteBoardScreen> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(),
-      body: Column(
-        children: [
-          SingleChildScrollView(
-            child: QuillToolbar.simple(
-              configurations: QuillSimpleToolbarConfigurations(
-                  controller: _controller,
-                  sharedConfigurations:
-                      const QuillSharedConfigurations(locale: Locale('ko'))),
+      appBar: AppBar(actions: [
+        TextButton(
+            onPressed: () {
+              if (_formKey.currentState!.validate()) {
+                ref.read(boardsProvider.notifier).addBoard(Board(
+                    _textEditingController.text,
+                    contentController.text,
+                    createdBy: '김동욱'));
+                Navigator.pop(context);
+              }
+            },
+            child: const Text('저장')),
+      ],),
+      body: SingleChildScrollView(
+        child: Column(
+          children: [
+            Form(
+              key: _formKey,
+              child: Column(
+                children: [
+                  TextFormField(
+                    controller: _textEditingController,
+                    validator: (value) {
+                      if (value == null || value.isEmpty) {
+                        return '제목을 입력해주세요.';
+                      }
+                      return null;
+                    },
+                  ),
+                  SizedBox(
+                    height: MediaQuery.of(context).size.height * 1.0,
+                    width: MediaQuery.of(context).size.width * 1.0,
+                    child: TextField(
+                      keyboardType: TextInputType.multiline,
+                      decoration: const InputDecoration(border: InputBorder.none),
+                      maxLines: 100,
+                      controller: contentController,
+                    ),
+                  ),
+                ],
+              ),
             ),
-          ),
-          Form(
-            key: _formKey,
-            child: Column(
-              children: [
-                TextFormField(
-                  controller: _textEditingController,
-                  validator: (value) {
-                    if (value == null || value.isEmpty) {
-                      return '제목을 입력해주세요.';
-                    }
-                    return null;
-                  },
-                ),
-              ],
-            ),
-          ),
-          Expanded(
-            child: QuillEditor.basic(
-              configurations: QuillEditorConfigurations(
-                  controller: _controller,
-                  readOnly: false,
-                  sharedConfigurations:
-                      const QuillSharedConfigurations(locale: Locale('ko'))),
-            ),
-          ),
-          TextButton(
-              onPressed: () {
-                if (_formKey.currentState!.validate()) {
-                  ref.read(boardsProvider.notifier).addBoard(Board(_textEditingController.text, jsonEncode(_controller.document.toDelta().toJson()), createdBy: '김동욱'));
-                  Navigator.pop(context);
-                }
-              },
-              child: const Text('저장')),
-        ],
+
+          ],
+        ),
       ),
     );
   }
@@ -81,7 +76,7 @@ class _WriteBoardScreenState extends ConsumerState<WriteBoardScreen> {
   @override
   void dispose() {
     _textEditingController.dispose();
-    _controller.dispose();
+    contentController.dispose();
     super.dispose();
   }
 }
