+++ lib/screen/main/tab/board/w_board_item.dart
@@ -1,10 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_quill/flutter_quill.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:playground/screen/main/tab/board/screen/s_board_detail.dart';
 import 'package:playground/screen/main/tab/board/vo/vo_board.dart';
-import 'package:playground/screen/main/tab/board/vo/vo_board_comment.dart';
-import 'package:playground/screen/main/tab/board/vo/vo_board_list.dart';
-import 'package:playground/screen/main/tab/board/w_board_detail.dart';
 
 class BoardItem extends ConsumerWidget {
   final Board _board;
