+++ pubspec.yaml
@@ -45,7 +45,7 @@ dependencies:
   device_info_plus: ^9.1.1
   super_clipboard: ^0.8.2+1
   table_calendar: ^3.0.9
-
+  flutter_markdown: ^0.6.22+1
   # data
   shared_preferences: ^2.2.2
   isar: ^3.1.0+1
